third_party/blink/web_tests/http/tests/security/data-url-inline.css.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<p>This test ensures that a data URL can access its own inline style sheets.
Sorry for the obscurity of the test case, but it's the repro from
<a href="https://bugs.webkit.org/show_bug.cgi?id=32309">Bug 32309</a>, which
has an "unobfuscated" version of the code.</p>
<p>This test passes if it alerts that it found 1 rule.</p>
<iframe src="data:text/html;base64,PGh0bWw+CiAgPHN0eWxlPgogICAgYm9keSB7CiAgICAgIGJhY2tncm91bmQ6IGdyZWVuOwogICAgfQogIDwvc3R5bGU+CiAgPGJvZHk+CiAgICA8c2NyaXB0PgogICAgICB2YXIgc3R5bGVzaGVldHMgPSBkb2N1bWVudC5zdHlsZVNoZWV0czsKICAgICAgdmFyIHN0eWxlc2hlZXQgPSBzdHlsZXNoZWV0c1tzdHlsZXNoZWV0cy5sZW5ndGgtMV07CiAgICAgIHZhciBydWxlcyA9IHN0eWxlc2hlZXQuY3NzUnVsZXM7CiAgICAgIGNvbnNvbGUubG9nKHJ1bGVzLmxlbmd0aCArICcgcnVsZXMgZm91bmQnKTsKICAgIDwvc2NyaXB0PgogIDwvYm9keT4KPC9odG1sPg=="></iframe>
